+++ hn/manager.cpp
@@ -22,10 +22,10 @@ namespace hhal {
     HNManagerExitCode HNManager::read_from_memory(int buffer_id, void *dest, size_t size) {
         return HNManagerExitCode::OK;
     }
-    HNManagerExitCode HNManager::write_sync_register(int event_id, uint8_t data) {
+    HNManagerExitCode HNManager::write_sync_register(int event_id, uint32_t data) {
         return HNManagerExitCode::OK;
     }
-    HNManagerExitCode HNManager::read_sync_register(int event_id, uint8_t *data) {
+    HNManagerExitCode HNManager::read_sync_register(int event_id, uint32_t *data) {
         return HNManagerExitCode::OK;
     }
 }
